+++ src/api/routes/hooks/transactions.js
@@ -1,13 +1,12 @@
 export default async (req, res, next) => {
   const shippoClientService = req.scope.resolve("shippoClientService")
-  const shippoWebhookService = req.scope.resolve("shippoWebhookService")
+  const shippoFulfillmentService = req.scope.resolve("shippoFulfillmentService")
   const eventBus = req.scope.resolve("eventBusService")
 
-  const validToken = await shippoWebhookService.verifyHookSecret(
+  const validToken = await shippoFulfillmentService.verifyHookSecret(
     req.query.token
   )
   const event = req.headers["x-shippo-event"]
-  // eventBus.emit(`shippo.received.${event}`, {})
 
   const invalidRequest = () => {
     eventBus.emit(`shippo.rejected.${event}`, {})
@@ -17,12 +16,6 @@ export default async (req, res, next) => {
 
   if (validToken) {
     const untrustedData = req.body.data
-
-    console.log(
-      "*********untrustedData: ",
-      JSON.stringify(untrustedData, null, 2)
-    )
-
     if (
       untrustedData.object_state === "VALID" &&
       untrustedData.status === "SUCCESS"
