+++ src/services/shippo-fulfillment.js
@@ -1,5 +1,6 @@
 import { FulfillmentService } from "medusa-interfaces"
-import { MedusaError } from "medusa-core-utils"
+import path from "path"
+import { getConfigFile, MedusaError } from "medusa-core-utils"
 import { shippoLineItem, shippoOrder } from "../utils/formatters"
 
 class ShippoFulfillmentService extends FulfillmentService {
@@ -20,6 +21,8 @@ class ShippoFulfillmentService extends FulfillmentService {
     options
   ) {
     super()
+
+    this.#setConfig(options)
 
     /** @private @const {CartService} */
     this.cartService_ = cartService
@@ -167,6 +170,12 @@ class ShippoFulfillmentService extends FulfillmentService {
     return true
   }
 
+  async verifyHookSecret(token) {
+    return this.options_.webhook_secret
+      ? this.options_.webhook_secret === token
+      : false
+  }
+
   async #createShippoOrder(order, fromAddress, lineItems, parcelName) {
     const client = this.#shippo.getClient()
     const params = await shippoOrder(order, fromAddress, lineItems, parcelName)
@@ -288,6 +297,17 @@ class ShippoFulfillmentService extends FulfillmentService {
         return service
       })
     )
+  }
+
+  #setConfig(options) {
+    if (Object.keys(options).length === 0) {
+      const {
+        configModule: { projectConfig },
+      } = getConfigFile(path.resolve("."), "medusa-config")
+      this.options_ = projectConfig
+    } else {
+      this.options_ = options
+    }
   }
 }
